Extensive changes in gui/ to make notationstaff a subclass of linedstaff. Not yet complete -- mostly it works, still some quirks to deal with.
Some real-time stuff in Composition and SegmentPerformanceHelper


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@1316 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/notationhlayout.cpp
@@ -1185,10 +1185,10 @@ NotationHLayout::isBarLineCorrect(StaffType &staff, unsigned int i)
 }
 
 Event *NotationHLayout::getTimeSignatureInBar(StaffType &staff,
-					      unsigned int i, int &timeSigX)
+					      unsigned int i, double &timeSigX)
 {
     BarData &bd(getBarData(staff)[i]);
-    timeSigX = bd.timeSigX;
+    timeSigX = (double)bd.timeSigX;
     return bd.timeSignature;
 }
